make compile. At 90%. Need to refactor UserInput.
