+++ lib/Config/Tonuino_config.h
@@ -1,13 +1,12 @@
-#ifndef DEFINES_H
-#define DEFINES_H
-#include <DFMiniMp3_interface.h>
+#ifndef TONUINO_CONFIG_H
+#define TONUINO_CONFIG_H
 
-//#define F_CPU 8000000UL
+#include <DFMiniMp3_interface.h>
 
 //DEBUG OUTPUT [just write "false" to disable debug output on serial]
 #define DEBUGSERIAL false
 
-//mp3 Messages -----------------------
+//Tonuino mp3 prompt IDs -----------------------
 #define MSG_DELETETAG 800
 #define MSG_ABORTEED 802
 #define MSG_ERROR 401
@@ -19,32 +18,6 @@
 #define MSG_SELECTFILE 320
 #define MSG_HELP 330
 #define MSG_CONFIRMED 340
-
-//Pin defines ------------------------
-#define PINANALOG_RNDMGEN A1 //ADC1 24
-// MFRC522
-#define RST_PIN 9 // PB1 15
-#define SS_PIN 10 // PB2 16
-//MP3
-#define DFMINI_BUSY 4 // PD4 6
-#define DFMINI_RX 2   // PD2 4
-#define DFMINI_TX 3   // PD3 5
-//Serial
-// RX 0  PD0 2
-// TX 1  PD1 3
-//HMI inputs
-//   #define ENCA 18 // PC5 27
-//   #define ENCB 17 // PC4 26
-//   #define ENCSW 19 // PC6 28
-#define PINPLPS 18 // PC5 27
-#define PINPREV 17 // PC4 26
-#define PINNEXT 19 // PC6 28
-//KeepAlive
-#define KEEPALIVE_PIN A0 //PB6 9
-#define KEEPALIVE_ACTIVE_STATE LOW
-//Feedback LED
-#define LED_PIN 13 //for test purposes only, otherwise 16 //PC0 25
-#define LED_ACTIVE_STATE HIGH
 
 //StatusLEDs interaction outputs
 #define FLASHSLOWMS 500
@@ -68,4 +41,4 @@
 #define WAIT_DFMINI_READY 1000 // 1 sec
 #define TIMEOUT_PROMPT_PLAYED 10000 // 10 sec
 
-#endif
+#endif // TONUINO_CONFIG_H
