+++ lib/DFMiniMp3_implementation/DFMiniMp3_implementation.h
@@ -1,6 +1,7 @@
 #ifndef DFMINIMP3_IMPLEMENTATION_H
 #define DFMINIMP3_IMPLEMENTATION_H
 #include <DFMiniMp3.h>
+#include <Arduino_config.h>
 #include <SoftwareSerial.h>
 #include <DfMiniMp3_interface.h>
