+++ lib/NfcTag/NfcTag.h
@@ -5,7 +5,7 @@
 #include <MFRC522_interface.h>
 //#include <SPI.h>
 #include <Folder.h>
-#include <Defines.h>
+//#include <Defines.h>
 
 // this object stores nfc tag data
 class NfcTag
