+++ platformio.ini
@@ -17,28 +17,28 @@ build_flags=
   -pthread
 lib_deps = 
   googletest
-  ArduinoFake # to be deleted!
-  Defines
+  Tonuino_config
   EEPROM_interface
-  Arduino_types
-  Folder
+  Arduino_interface
   MFRC522_interface
+  #PowerManager/PowerManager_interface
+  Folder
   NfcTag
   Mp3PlayerControl
 # Libs that are not needed for unit testing on native
 lib_ignore =
-  Defines
   ClickEncoder
   DFMiniMp3
   EEPROM
   EEPROM_implementation # TODO: ignore as ArduinoFake currently does not yield SPI code that this is depending on!
+  Tonuino/InputManager
+  PowerManager/PowerManager
   interface_UserInput
   MFRC522
   MFRC522_implementation
   SoftwareSerial
   SPI
-  PowerManager
-
+  
 # TODO: exclude libs from build that are out of scope for unittest
 lib_compat_mode = off
